requirements.txt
requests==2.31.0
# Whisper (OpenAI 공식 저장소에서 설치)
git+https://github.com/openai/whisper.git

# ffmpeg-python: Whisper 내부에서 ffmpeg 호출에 사용됨
ffmpeg-python

# numpy 등 whisper 내부에서 사용하는 기본 패키지는 whisper 설치 시 자동 설치됨

# pyannote.audio의 의존 패키지
torch==2.1.2
torchaudio==2.1.2
numpy
scipy
librosa
transformers>=4.31.0
huggingface_hub
tqdm
openai-whisper
pydub